Membuat API untuk login dan juga logout, tambahan membuat fitur update untuk user menggunakan verifikasi old password

+++ routes/api.php
@@ -1,21 +1,30 @@
 <?php
 
-use App\Http\Controllers\CategoryController;
-use App\Http\Controllers\TodoController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\TodoController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\CategoryController;
 
 // Route::get('/user', function (Request $request) {
 //     return $request->user();
 // });
 
 Route::get('/users', [UserController::class, 'index']);
+Route::patch('/user/{id}', [UserController::class, 'update']);
+
+
 Route::get('/categories', [CategoryController::class, 'index']);
 
 
 Route::get('/todos', [TodoController::class, 'index']);
 Route::get('/todo/{id}', [TodoController::class, 'show']);
-Route::post('/todo', [TodoController::class, 'store']);
-Route::patch('/todo/{id}', [TodoController::class, 'update']);
-Route::delete('/todo/{id}', [TodoController::class, 'destroy']);
+Route::post('/todo', [TodoController::class, 'store'])->middleware('auth:sanctum');
+Route::patch('/todo/{id}', [TodoController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/todo/{id}', [TodoController::class, 'destroy'])->middleware('auth:sanctum');
+
+Route::prefix('auth')->group(function() {
+    Route::post('/login', [AuthController::class, 'login']);
+    Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
+});
